PIM-5384: Move import method in proper class
